+++ Cookie.API/Protocol/Network/Messages/Debug/DebugHighlightCellsMessage.cs
@@ -7,7 +7,7 @@ namespace Cookie.API.Protocol.Network.Messages.Debug
     {
         public const ushort ProtocolId = 2001;
 
-        public DebugHighlightCellsMessage(int color, List<ushort> cells)
+        public DebugHighlightCellsMessage(double color, List<ushort> cells)
         {
             Color = color;
             Cells = cells;
@@ -18,12 +18,12 @@ namespace Cookie.API.Protocol.Network.Messages.Debug
         }
 
         public override ushort MessageID => ProtocolId;
-        public int Color { get; set; }
+        public double Color { get; set; }
         public List<ushort> Cells { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(Color);
+            writer.WriteDouble(Color);
             writer.WriteShort((short) Cells.Count);
             for (var cellsIndex = 0; cellsIndex < Cells.Count; cellsIndex++)
                 writer.WriteVarUhShort(Cells[cellsIndex]);
@@ -31,7 +31,7 @@ namespace Cookie.API.Protocol.Network.Messages.Debug
 
         public override void Deserialize(IDataReader reader)
         {
-            Color = reader.ReadInt();
+            Color = reader.ReadDouble();
             var cellsCount = reader.ReadUShort();
             Cells = new List<ushort>();
             for (var cellsIndex = 0; cellsIndex < cellsCount; cellsIndex++)
