+++ Cookie.API/Protocol/Network/Messages/Game/Context/GameMapMovementRequestMessage.cs
@@ -7,7 +7,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context
     {
         public const ushort ProtocolId = 950;
 
-        public GameMapMovementRequestMessage(List<short> keyMovements, int mapId)
+        public GameMapMovementRequestMessage(List<short> keyMovements, double mapId)
         {
             KeyMovements = keyMovements;
             MapId = mapId;
@@ -19,14 +19,14 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context
 
         public override ushort MessageID => ProtocolId;
         public List<short> KeyMovements { get; set; }
-        public int MapId { get; set; }
+        public double MapId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
             writer.WriteShort((short) KeyMovements.Count);
             for (var keyMovementsIndex = 0; keyMovementsIndex < KeyMovements.Count; keyMovementsIndex++)
                 writer.WriteShort(KeyMovements[keyMovementsIndex]);
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
         }
 
         public override void Deserialize(IDataReader reader)
@@ -35,7 +35,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context
             KeyMovements = new List<short>();
             for (var keyMovementsIndex = 0; keyMovementsIndex < keyMovementsCount; keyMovementsIndex++)
                 KeyMovements.Add(reader.ReadShort());
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
         }
     }
 }
