+++ Cookie.API/Protocol/Network/Messages/Game/Context/Mount/PaddockBuyResultMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Mount
     {
         public const ushort ProtocolId = 6516;
 
-        public PaddockBuyResultMessage(int paddockId, bool bought, ulong realPrice)
+        public PaddockBuyResultMessage(double paddockId, bool bought, ulong realPrice)
         {
             PaddockId = paddockId;
             Bought = bought;
@@ -18,20 +18,20 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Mount
         }
 
         public override ushort MessageID => ProtocolId;
-        public int PaddockId { get; set; }
+        public double PaddockId { get; set; }
         public bool Bought { get; set; }
         public ulong RealPrice { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(PaddockId);
+            writer.WriteDouble(PaddockId);
             writer.WriteBoolean(Bought);
             writer.WriteVarUhLong(RealPrice);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            PaddockId = reader.ReadInt();
+            PaddockId = reader.ReadDouble();
             Bought = reader.ReadBoolean();
             RealPrice = reader.ReadVarUhLong();
         }
