+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/ChangeMapMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay
     {
         public const ushort ProtocolId = 221;
 
-        public ChangeMapMessage(int mapId)
+        public ChangeMapMessage(double mapId)
         {
             MapId = mapId;
         }
@@ -16,16 +16,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay
         }
 
         public override ushort MessageID => ProtocolId;
-        public int MapId { get; set; }
+        public double MapId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
         }
     }
 }
