+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/CurrentMapMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay
     {
         public const ushort ProtocolId = 220;
 
-        public CurrentMapMessage(int mapId, string mapKey)
+        public CurrentMapMessage(double mapId, string mapKey)
         {
             MapId = mapId;
             MapKey = mapKey;
@@ -17,18 +17,18 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay
         }
 
         public override ushort MessageID => ProtocolId;
-        public int MapId { get; set; }
+        public double MapId { get; set; }
         public string MapKey { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
             writer.WriteUTF(MapKey);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
             MapKey = reader.ReadUTF();
         }
     }
