+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/Death/GameRolePlayPlayerLifeStatusMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Death
     {
         public const ushort ProtocolId = 5996;
 
-        public GameRolePlayPlayerLifeStatusMessage(byte state, int phenixMapId)
+        public GameRolePlayPlayerLifeStatusMessage(byte state, double phenixMapId)
         {
             State = state;
             PhenixMapId = phenixMapId;
@@ -18,18 +18,18 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Death
 
         public override ushort MessageID => ProtocolId;
         public byte State { get; set; }
-        public int PhenixMapId { get; set; }
+        public double PhenixMapId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
             writer.WriteByte(State);
-            writer.WriteInt(PhenixMapId);
+            writer.WriteDouble(PhenixMapId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
             State = reader.ReadByte();
-            PhenixMapId = reader.ReadInt();
+            PhenixMapId = reader.ReadDouble();
         }
     }
 }
