+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/Lockable/LockableStateUpdateStorageMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Lockable
     {
         public new const ushort ProtocolId = 5669;
 
-        public LockableStateUpdateStorageMessage(int mapId, uint elementId)
+        public LockableStateUpdateStorageMessage(double mapId, uint elementId)
         {
             MapId = mapId;
             ElementId = elementId;
@@ -17,20 +17,20 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Lockable
         }
 
         public override ushort MessageID => ProtocolId;
-        public int MapId { get; set; }
+        public double MapId { get; set; }
         public uint ElementId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
             base.Serialize(writer);
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
             writer.WriteVarUhInt(ElementId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
             base.Deserialize(reader);
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
             ElementId = reader.ReadVarUhInt();
         }
     }
