+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/Npc/NpcDialogCreationMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Npc
     {
         public const ushort ProtocolId = 5618;
 
-        public NpcDialogCreationMessage(int mapId, int npcId)
+        public NpcDialogCreationMessage(double mapId, int npcId)
         {
             MapId = mapId;
             NpcId = npcId;
@@ -17,18 +17,18 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Npc
         }
 
         public override ushort MessageID => ProtocolId;
-        public int MapId { get; set; }
+        public double MapId { get; set; }
         public int NpcId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
             writer.WriteInt(NpcId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
             NpcId = reader.ReadInt();
         }
     }
