+++ Cookie.API/Protocol/Network/Messages/Game/Context/Roleplay/Npc/NpcGenericActionRequestMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Npc
     {
         public const ushort ProtocolId = 5898;
 
-        public NpcGenericActionRequestMessage(int npcId, byte npcActionId, int npcMapId)
+        public NpcGenericActionRequestMessage(int npcId, byte npcActionId, double npcMapId)
         {
             NpcId = npcId;
             NpcActionId = npcActionId;
@@ -20,20 +20,20 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Context.Roleplay.Npc
         public override ushort MessageID => ProtocolId;
         public int NpcId { get; set; }
         public byte NpcActionId { get; set; }
-        public int NpcMapId { get; set; }
+        public double NpcMapId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
             writer.WriteInt(NpcId);
             writer.WriteByte(NpcActionId);
-            writer.WriteInt(NpcMapId);
+            writer.WriteDouble(NpcMapId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
             NpcId = reader.ReadInt();
             NpcActionId = reader.ReadByte();
-            NpcMapId = reader.ReadInt();
+            NpcMapId = reader.ReadDouble();
         }
     }
 }
