+++ Cookie.API/Protocol/Network/Messages/Game/Guild/GuildPaddockRemovedMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild
     {
         public const ushort ProtocolId = 5955;
 
-        public GuildPaddockRemovedMessage(int paddockId)
+        public GuildPaddockRemovedMessage(double paddockId)
         {
             PaddockId = paddockId;
         }
@@ -16,16 +16,16 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild
         }
 
         public override ushort MessageID => ProtocolId;
-        public int PaddockId { get; set; }
+        public double PaddockId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(PaddockId);
+            writer.WriteDouble(PaddockId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            PaddockId = reader.ReadInt();
+            PaddockId = reader.ReadDouble();
         }
     }
 }
