+++ Cookie.API/Protocol/Network/Messages/Game/Guild/Tax/GuildFightLeaveRequestMessage.cs
@@ -6,7 +6,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
     {
         public const ushort ProtocolId = 5715;
 
-        public GuildFightLeaveRequestMessage(int taxCollectorId, ulong characterId)
+        public GuildFightLeaveRequestMessage(double taxCollectorId, ulong characterId)
         {
             TaxCollectorId = taxCollectorId;
             CharacterId = characterId;
@@ -17,18 +17,18 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
         }
 
         public override ushort MessageID => ProtocolId;
-        public int TaxCollectorId { get; set; }
+        public double TaxCollectorId { get; set; }
         public ulong CharacterId { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(TaxCollectorId);
+            writer.WriteDouble(TaxCollectorId);
             writer.WriteVarUhLong(CharacterId);
         }
 
         public override void Deserialize(IDataReader reader)
         {
-            TaxCollectorId = reader.ReadInt();
+            TaxCollectorId = reader.ReadDouble();
             CharacterId = reader.ReadVarUhLong();
         }
     }
