+++ Cookie.API/Protocol/Network/Messages/Game/Guild/Tax/GuildFightPlayersEnemiesListMessage.cs
@@ -8,7 +8,8 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
     {
         public const ushort ProtocolId = 5928;
 
-        public GuildFightPlayersEnemiesListMessage(int fightId, List<CharacterMinimalPlusLookInformations> playerInfo)
+        public GuildFightPlayersEnemiesListMessage(double fightId,
+            List<CharacterMinimalPlusLookInformations> playerInfo)
         {
             FightId = fightId;
             PlayerInfo = playerInfo;
@@ -19,12 +20,12 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
         }
 
         public override ushort MessageID => ProtocolId;
-        public int FightId { get; set; }
+        public double FightId { get; set; }
         public List<CharacterMinimalPlusLookInformations> PlayerInfo { get; set; }
 
         public override void Serialize(IDataWriter writer)
         {
-            writer.WriteInt(FightId);
+            writer.WriteDouble(FightId);
             writer.WriteShort((short) PlayerInfo.Count);
             for (var playerInfoIndex = 0; playerInfoIndex < PlayerInfo.Count; playerInfoIndex++)
             {
@@ -35,7 +36,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
 
         public override void Deserialize(IDataReader reader)
         {
-            FightId = reader.ReadInt();
+            FightId = reader.ReadDouble();
             var playerInfoCount = reader.ReadUShort();
             PlayerInfo = new List<CharacterMinimalPlusLookInformations>();
             for (var playerInfoIndex = 0; playerInfoIndex < playerInfoCount; playerInfoIndex++)
