+++ Cookie.API/Protocol/Network/Messages/Game/Guild/Tax/TaxCollectorAttackedMessage.cs
@@ -7,8 +7,8 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
     {
         public const ushort ProtocolId = 5918;
 
-        public TaxCollectorAttackedMessage(ushort firstNameId, ushort lastNameId, short worldX, short worldY, int mapId,
-            ushort subAreaId, BasicGuildInformations guild)
+        public TaxCollectorAttackedMessage(ushort firstNameId, ushort lastNameId, short worldX, short worldY,
+            double mapId, ushort subAreaId, BasicGuildInformations guild)
         {
             FirstNameId = firstNameId;
             LastNameId = lastNameId;
@@ -28,7 +28,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
         public ushort LastNameId { get; set; }
         public short WorldX { get; set; }
         public short WorldY { get; set; }
-        public int MapId { get; set; }
+        public double MapId { get; set; }
         public ushort SubAreaId { get; set; }
         public BasicGuildInformations Guild { get; set; }
 
@@ -38,7 +38,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
             writer.WriteVarUhShort(LastNameId);
             writer.WriteShort(WorldX);
             writer.WriteShort(WorldY);
-            writer.WriteInt(MapId);
+            writer.WriteDouble(MapId);
             writer.WriteVarUhShort(SubAreaId);
             Guild.Serialize(writer);
         }
@@ -49,7 +49,7 @@ namespace Cookie.API.Protocol.Network.Messages.Game.Guild.Tax
             LastNameId = reader.ReadVarUhShort();
             WorldX = reader.ReadShort();
             WorldY = reader.ReadShort();
-            MapId = reader.ReadInt();
+            MapId = reader.ReadDouble();
             SubAreaId = reader.ReadVarUhShort();
             Guild = new BasicGuildInformations();
             Guild.Deserialize(reader);
